% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_knockoff_stan.R
\name{categorical_knockoff_stan}
\alias{categorical_knockoff_stan}
\title{Use Bayesian method to generate knockoff copy for data with categorical covariates.}
\usage{
categorical_knockoff_stan(dataX, burnin = 1000)
}
\arguments{
\item{dataX}{Predictor dataframe, rows correspond to observations, columns correspond to covariates.}

\item{burnin}{The number of MCMC burnin for stan}
}
\value{
A dataframe with generated knockoff copy
}
\description{
Use Bayesian method to generate knockoff copy for data with categorical covariates.
}
