% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_surv.R
\name{categorical_surv}
\alias{categorical_surv}
\title{Select features for data with categorical covariates and survival outcome using knockoff filter with group lasso.}
\usage{
categorical_surv(dataX, knockoffCopy, y, fdr = 0.2, offset = 0)
}
\arguments{
\item{dataX}{The original data frame, where rows correspond to observations, while columns correspond to predictors. Categorical covariates need to be factors.}

\item{knockoffCopy}{The Knockoff copy containing categorical covariates. Can be produced by the 'categorical_knockoff_stan' of this package.}

\item{y}{The 'Surv' type right-censored survival outcome}

\item{fdr}{The target false discovery rate}

\item{offset}{Offset for the group knockoff}
}
\value{
The name of the predictors selected under the target FDR
}
\description{
Select features for data with categorical covariates and survival outcome using knockoff filter with group lasso.
}
\examples{
library(BayesianKnockoff)
library(survival)
set.seed(1)
data0<-data.frame(matrix(sample(1:3,1000,replace = TRUE),ncol=10))
data1<-apply(data0,2,as.character)
data1 <- as.data.frame(unclass(data1),stringsAsFactors = TRUE)
data2<-data.frame(matrix(rnorm(1000),ncol=10))
colnames(data2)<-paste0("X",11:20)
dataX<-cbind(data1,data2)
knockoffCopy<-categorical_knockoff_stan(dataX)

time0<-(data0$X5-2)*10-(data0$X9-2)*10+data2$X11*10+data2$X12*10
time0<-time0-min(time0)+1
time1<-rexp(nrow(dataX),0.01)
cens<-time0<time1
time<-ifelse(cens,time0,time1)
tempsurv<-categorical_surv(dataX,knockoffCopy,Surv(time,cens))
}
