% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_knockoff_stan.R
\name{microbiome_knockoff_stan}
\alias{microbiome_knockoff_stan}
\title{Produce knockoff copy for microbiome data}
\usage{
microbiome_knockoff_stan(microbiomeCount, burnin = 1000)
}
\arguments{
\item{microbiomeCount}{A matrix if microbiome sequence counts, rows correspond to observations, columns correspond to taxa.}

\item{burnin}{The number of MCMC burnin for stan}
}
\value{
knockoffCopy - Generate knockoff copy for microbiome data
}
\description{
Produce knockoff copy for microbiome data
}
