% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_select.R
\name{microbiome_select}
\alias{microbiome_select}
\title{Run feature selection for microbiome data}
\usage{
microbiome_select(
  microbiomeCount,
  knockoffCopy,
  y,
  Tree = NULL,
  fdr = 0.2,
  offset = 0
)
}
\arguments{
\item{microbiomeCount}{A matrix of the original microbiome sequence counts, rows correspond to observations, columns correspond to taxa. Users must provide column names.}

\item{knockoffCopy}{A matrix of knockoff microbiome sequence counts.}

\item{y}{A continuous outcome}

\item{Tree}{A matrix reflecting the phylogenetic relationships between taxa. The rows correspond to the tip node of the phylogenetic tree, while the columns correspond to the tip and internal nodes of the phylogenetic tree. For each row(tip node), the entries that correspond  to the phylogenentic tree node (including the tip node)that locate on the path from the root of the tree to the tip node are 1s, while other entries are 0s.}

\item{fdr}{Target fdr}

\item{offset}{The offset for knockoff+}
}
\value{
A list with selected taxa or tree node (if phylogenetic tree is provided)
\itemize{
\item selected - The taxa selected
\item nodeSelected - If phylogenetic tree is provided, the node selected will also be provided
}
}
\description{
Run feature selection for microbiome data
}
\examples{
\dontrun{
library(BayesianKnockoff)
data("HIVdata")
knockoffCopy<-microbiome_knockoff_stan(HIVdata$microbiomeCount) #may take a few hours
temp1<-microbiome_select(HIVdata$microbiomeCount,knockoffCopy,HIVdata$y)
temp2<-microbiome_select(HIVdata$microbiomeCount,knockoffCopy,HIVdata$y,HIVdata$Tree)
}
}
\references{
Bien, J., Yan, X., Simpson, L. et al. Tree-aggregated predictive modeling of microbiome data. Sci Rep 11, 14505 (2021). \url{https://doi.org/10.1038/s41598-021-93645-3}
}
